Refactor com_atproto::server
