+++ atprotolib-rs/src/types/com_atproto/mod.rs
@@ -13,61 +13,7 @@ pub mod moderation;
 pub mod repo;
 
 /// Contains types for the `com.atproto.server` namespace.
-pub mod server {
-    mod check_account_status;
-    mod confirm_email;
-    mod create_account;
-    mod create_app_password;
-    mod create_invite_code;
-    mod create_invite_codes;
-    mod create_session;
-    mod deactivate_account;
-    mod defs;
-    mod delete_account;
-    mod describe_server;
-    mod get_account_invite_codes;
-    mod get_service_auth;
-    mod get_session;
-    mod list_app_passwords;
-    mod refresh_session;
-    mod request_email_update;
-    mod request_password_reset;
-    mod reserve_signing_key;
-    mod reset_password;
-    mod revoke_app_password;
-    mod update_email;
-
-    pub use self::{
-        check_account_status::*,
-        confirm_email::*,
-        create_account::*,
-        create_app_password::*,
-        create_invite_code::*,
-        create_invite_codes::*,
-        create_session::*,
-        deactivate_account::*,
-        defs::*,
-        delete_account::*,
-        describe_server::*,
-        // get_account_invite_codes::*,
-        get_service_auth::*,
-        get_session::*,
-        list_app_passwords::*,
-        refresh_session::*,
-        request_email_update::*,
-        request_password_reset::*,
-        reserve_signing_key::*,
-        reset_password::*,
-        revoke_app_password::*,
-        update_email::*
-    };
-
-    #[cfg(feature = "apicalls")]
-    mod api_calls;
-
-    #[cfg(feature = "apicalls")]
-    pub use self::api_calls::*;
-}
+pub mod server;
 
 /// Contains types for the `com.atproto.sync` namespace.
 pub mod sync {
