--- atprotolib-rs/src/types/com_atproto/server/get_account_invite_codes.rs
@@ -1,15 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use super::InviteCode;
-
-/*
-    com.atproto.server.getAccountInviteCodes
-*/
-
-/// Represents a request to get account invite codes.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GetAccountInviteCodesRequest {
-    /// Codes to get.
-    #[serde(rename = "codes")]
-    pub codes: Vec<InviteCode>
-}
